requirements.txt
streamlit==1.28.0
opencv-python==4.8.1.78
insightface==0.7.3
onnx==1.16.1
onnxruntime==1.16.3
pymongo==4.5.0
python-dotenv==1.0.0
numpy==1.24.3
pillow==10.0.1
pandas==2.1.1
altair==5.1.2
python-dateutil==2.8.2
dnspython==2.4.2
streamlit-camera-input-live